+++ src/modifiedConstants/modifiedAlliances.js
